Timer, Registers, Voltage
Quite a few changes here.
Added support for monitoring voltages on a servo(or all) and cache the
values.

Changed Logical registers to be a union/Structure and combined both the
low ones and the high ones into same structure and changed the code to
use this structure.  Also added Timeouts for several things as
settable/gettable settings.

Timer code updated, I have a 16 bit timer and function to retrieve
current timer (interrupt safe), plus compare function, that I use in the
Voltage and Pose sections.  May change other 8 bit timeouts to use as
well.

Changed Readme to readme.md so markdown shows up on github